--- src/main/java/acme/features/authenticated/customer/booking/CustomerBookingShowService.java
@@ -1,56 +1,0 @@
-
-package acme.features.authenticated.customer.booking;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-import acme.client.components.models.Dataset;
-import acme.client.services.AbstractGuiService;
-import acme.client.services.GuiService;
-import acme.entities.booking.Booking;
-import acme.realms.customer.Customer;
-
-@GuiService
-public class CustomerBookingShowService extends AbstractGuiService<Customer, Booking> {
-
-	@Autowired
-	private CustomerBookingRepository repository;
-
-
-	@Override
-	public void authorise() {
-		boolean status;
-		int masterId;
-		Booking booking;
-		Customer customer;
-
-		masterId = super.getRequest().getData("id", int.class);
-		booking = this.repository.findBookingById(masterId);
-		customer = null;
-		if (booking != null)
-			customer = booking.getCustomer();
-		status = super.getRequest().getPrincipal().hasRealm(customer) && booking != null;
-
-		super.getResponse().setAuthorised(status);
-	}
-
-	@Override
-	public void load() {
-		Booking booking;
-		int id;
-
-		id = super.getRequest().getData("id", int.class);
-		booking = this.repository.findBookingById(id);
-
-		super.getBuffer().addData(booking);
-	}
-
-	@Override
-	public void unbind(final Booking booking) {
-		Dataset dataset;
-
-		dataset = super.unbindObject(booking, "locatorCode", "purchaseMoment", "travelClass", "price", "lastNibble", "draftMode");
-
-		super.getResponse().addData(dataset);
-	}
-
-}
