+++ src/main/java/acme/features/customer/booking/CustomerBookingCreateService.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -10,7 +10,7 @@ import acme.entities.booking.Booking;
 import acme.realms.customer.Customer;
 
 @GuiService
-public class CustomerBookingPublishService extends AbstractGuiService<Customer, Booking> {
+public class CustomerBookingCreateService extends AbstractGuiService<Customer, Booking> {
 
 	@Autowired
 	private CustomerBookingRepository repository;
@@ -18,28 +18,19 @@ public class CustomerBookingPublishService extends AbstractGuiService<Customer, 
 
 	@Override
 	public void authorise() {
-		boolean status;
-		int masterId;
-		Booking booking;
-		Customer customer;
-
-		masterId = super.getRequest().getData("id", int.class);
-		booking = this.repository.findBookingById(masterId);
-		customer = null;
-		if (booking != null)
-			customer = booking.getCustomer();
-		status = super.getRequest().getPrincipal().hasRealm(customer) && booking != null && booking.getDraftMode();
-
-		super.getResponse().setAuthorised(status);
+		super.getResponse().setAuthorised(true);
 	}
 
 	@Override
 	public void load() {
 		Booking booking;
-		int id;
+		Customer customer;
 
-		id = super.getRequest().getData("id", int.class);
-		booking = this.repository.findBookingById(id);
+		customer = (Customer) super.getRequest().getPrincipal().getActiveRealm();
+
+		booking = new Booking();
+		booking.setDraftMode(true);
+		booking.setCustomer(customer);
 
 		super.getBuffer().addData(booking);
 	}
@@ -58,21 +49,11 @@ public class CustomerBookingPublishService extends AbstractGuiService<Customer, 
 
 	@Override
 	public void validate(final Booking booking) {
-		boolean status;
-		boolean isLastNibbleNull;
-
-		isLastNibbleNull = false;
-		if (booking.getLastNibble() == null)
-			isLastNibbleNull = true;
-
-		status = isLastNibbleNull;
-
-		super.state(status, "*", "customer.booking.publish.lastNibble-null");
+		;
 	}
 
 	@Override
 	public void perform(final Booking booking) {
-		booking.setDraftMode(false);
 		this.repository.save(booking);
 	}
