+++ src/main/java/acme/features/customer/booking/CustomerBookingListService.java
@@ -1,19 +1,18 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
 import acme.client.components.models.Dataset;
-import acme.client.services.AbstractService;
+import acme.client.services.AbstractGuiService;
 import acme.client.services.GuiService;
 import acme.entities.booking.Booking;
 import acme.realms.customer.Customer;
 
 @GuiService
-public class CustomerBookingListingService extends AbstractService<Customer, Booking> {
+public class CustomerBookingListService extends AbstractGuiService<Customer, Booking> {
 
 	@Autowired
 	private CustomerBookingRepository repository;
@@ -26,10 +25,10 @@ public class CustomerBookingListingService extends AbstractService<Customer, Boo
 
 	@Override
 	public void load() {
-		List<Booking> bookings = new ArrayList<>();
+		Collection<Booking> bookings;
 		int customerId;
 
-		customerId = super.getRequest().getPrincipal().getAccountId();
+		customerId = super.getRequest().getPrincipal().getActiveRealm().getId();
 		bookings = this.repository.findBookingsByCustomerId(customerId);
 
 		super.getBuffer().addData(bookings);
@@ -40,7 +39,7 @@ public class CustomerBookingListingService extends AbstractService<Customer, Boo
 	public void unbind(final Booking booking) {
 		Dataset dataset;
 
-		dataset = super.unbindObject(booking, "locatorCode", "travelClass", "price");
+		dataset = super.unbindObject(booking, "locatorCode", "travelClass", "price", "draftMode");
 		super.getResponse().addData(dataset);
 	}
