+++ src/main/java/acme/features/customer/booking/CustomerBookingUpdateService.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -10,7 +10,7 @@ import acme.entities.booking.Booking;
 import acme.realms.customer.Customer;
 
 @GuiService
-public class CustomerBookingCreateService extends AbstractGuiService<Customer, Booking> {
+public class CustomerBookingUpdateService extends AbstractGuiService<Customer, Booking> {
 
 	@Autowired
 	private CustomerBookingRepository repository;
@@ -18,19 +18,28 @@ public class CustomerBookingCreateService extends AbstractGuiService<Customer, B
 
 	@Override
 	public void authorise() {
-		super.getResponse().setAuthorised(true);
+		boolean status;
+		int masterId;
+		Booking booking;
+		Customer customer;
+
+		masterId = super.getRequest().getData("id", int.class);
+		booking = this.repository.findBookingById(masterId);
+		customer = null;
+		if (booking != null)
+			customer = booking.getCustomer();
+		status = super.getRequest().getPrincipal().hasRealm(customer) && booking != null && !booking.getDraftMode();
+
+		super.getResponse().setAuthorised(status);
 	}
 
 	@Override
 	public void load() {
 		Booking booking;
-		Customer customer;
+		int id;
 
-		customer = (Customer) super.getRequest().getPrincipal().getActiveRealm();
-
-		booking = new Booking();
-		booking.setDraftMode(true);
-		booking.setCustomer(customer);
+		id = super.getRequest().getData("id", int.class);
+		booking = this.repository.findBookingById(id);
 
 		super.getBuffer().addData(booking);
 	}
@@ -65,4 +74,5 @@ public class CustomerBookingCreateService extends AbstractGuiService<Customer, B
 
 		super.getResponse().addData(dataset);
 	}
+
 }
