+++ src/main/resources/WEB-INF/views/customer/booking/form.jsp
@@ -1,22 +1,33 @@
+<%--
+- form.jsp
+-
+- Copyright (C) 2012-2025 Rafael Corchuelo.
+-
+- In keeping with the traditional purpose of furthering education and research, it is
+- the policy of the copyright owner to permit non-commercial use and redistribution of
+- this software. It has been tested carefully, but it is not guaranteed for any particular
+- purposes.  The copyright owner does not offer any warranties or representations, nor do
+- they accept any liabilities with respect to them.
+--%>
+
 <%@page%>
 
 <%@taglib prefix="jstl" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@taglib prefix="acme" uri="http://acme-framework.org/"%>
 
-<acme:form readonly="true">
+<acme:form>
 	<acme:input-textbox code="customer.booking.form.label.locatorCode" path="locatorCode"/>
 	<acme:input-moment code="customer.booking.form.label.purchaseMoment" path="purchaseMoment"/>	
 	<acme:input-textbox code="customer.booking.form.label.travelClass" path="travelClass"/>
 	<acme:input-money code="customer.booking.form.label.price" path="price"/>
 	<acme:input-textbox code="customer.booking.form.label.lastNibble" path="lastNibble"/>
-	<acme:list-column code="authenticated.customer.list.label.draftMode" path="draftMode"/>		
+	<acme:input-checkbox code="customer.booking.form.label.draftMode" path="draftMode"/>		
 	
 	<jstl:choose>	 
 		<jstl:when test="${_command == 'show' && draftMode == false}">
 			<acme:button code="customer.booking.form.button.passsengers" action="/customer/passsenger/list?masterId=${id}"/> <%-- Esto hay que revisarlo --%>	
 		</jstl:when>
-		<jstl:when test="${acme:anyOf(_command, 'show|update|delete|publish') && draftMode == true}">
-			<acme:button code="customer.booking.form.button.passsengers" action="/customer/passsenger/list?masterId=${id}"/> <%-- Esto hay que revisarlo --%>	
+		<jstl:when test="${acme:anyOf(_command, 'show|update|publish') && draftMode == true}">
 			<acme:submit code="customer.booking.form.button.update" action="/customer/booking/update"/>
 			<acme:submit code="customer.booking.form.button.publish" action="/customer/booking/publish"/>
 		</jstl:when>
