+++ src/main/resources/WEB-INF/views/fragments/menu.jsp
@@ -47,7 +47,7 @@
 			<acme:menu-suboption code="master.menu.consumer.favourite-link" action="http://www.example.com/"/>
 		</acme:menu-option>
 		
-		<acme:menu-option code="master.menu.manager" access="hasRealm('Customer')">
+		<acme:menu-option code="master.menu.customer" access="hasRealm('Customer')">
  			<acme:menu-suboption code="master.menu.customer.list-my-bookings" action="/customer/booking/list"/>
  		</acme:menu-option>
 	</acme:menu-left>
